refactor: refactor common presentation layer
