+++ common/src/presentation/AnomalyPresenter.ts
@@ -1,18 +1,18 @@
-import { Anomaly } from '../../../domain/core/Anomaly.js'
-import { Intrusion } from '../../../domain/core/Intrusion.js'
-import { Outlier } from '../../../domain/core/Outlier.js'
-import { Detection } from '../../../domain/core/Detection.js'
-import { MeasureType } from '../../../domain/core/MeasureType.js'
-import { MeasureUnit } from '../../../domain/core/MeasureUnit.js'
-import { ObjectClass } from '../../../domain/core/ObjectClass.js'
-import { AnomalyFactory } from '../../../domain/factories/AnomalyFactory.js'
-import { DetectionFactory } from '../../../domain/factories/DetectionFactory.js'
-import { MeasurementFactory } from '../../../domain/factories/MeasurementFactory.js'
-import { MeasureFactory } from '../../../domain/factories/MeasureFactory.js'
-import { AnomalySchema, anomalySchema } from '../../schemas/AnomalySchema.js'
-import { NumericMeasurement } from '../../../domain/core/NumericMeasurement.js'
+import { Anomaly } from '../domain/core/Anomaly.js'
+import { Intrusion } from '../domain/core/Intrusion.js'
+import { Outlier } from '../domain/core/Outlier.js'
+import { Detection } from '../domain/core/Detection.js'
+import { MeasureType } from '../domain/core/MeasureType.js'
+import { MeasureUnit } from '../domain/core/MeasureUnit.js'
+import { ObjectClass } from '../domain/core/ObjectClass.js'
+import { AnomalyFactory } from '../domain/factories/AnomalyFactory.js'
+import { DetectionFactory } from '../domain/factories/DetectionFactory.js'
+import { MeasurementFactory } from '../domain/factories/MeasurementFactory.js'
+import { MeasureFactory } from '../domain/factories/MeasureFactory.js'
+import { AnomalySchema, anomalySchema } from './schemas/AnomalySchema.js'
+import { NumericMeasurement } from '../domain/core/NumericMeasurement.js'
 
-export class AnomaliesAdapter {
+export class AnomalyPresenter {
   static asDomainEvent(anomalyObj: object): Anomaly {
     const anomalyMessage: AnomalySchema = anomalySchema.parse(anomalyObj)
     if (anomalyMessage.type === 'intrusion') {
