+++ common/src/presentation/DetectionPresenter.ts
@@ -1,7 +1,7 @@
-import { detectionSchema } from '../../schemas/DetectionSchema.js'
-import { Detection } from '../../../domain/core'
+import { detectionSchema } from './schemas/DetectionSchema.js'
+import { Detection } from '../domain/core'
 
-export class DetectionsAdapter {
+export class DetectionPresenter {
   static asDomainEvent(detectionObj: object): Detection {
     return detectionSchema.parse(detectionObj)
   }
