+++ common/src/presentation/MeasurementPresenter.ts
@@ -1,7 +1,7 @@
-import { Measurement } from '../../../domain/core'
-import { measurementSchema } from '../../schemas/MeasurementSchema.js'
+import { Measurement } from '../domain/core'
+import { measurementSchema } from './schemas/MeasurementSchema.js'
 
-export class MeasurementsAdapter {
+export class MeasurementPresenter {
   static asDomainEvent(measurementObj: object): Measurement {
     return measurementSchema.parse(measurementObj)
   }
