+++ location/src/infrastructure/api/routes/locationRouter.ts
@@ -4,6 +4,7 @@ import { locationsController } from '../controllers/locationController'
 import { LocationPresenter } from '@/presentation/api/LocationPresenter'
 import { ZodLocationPresenter } from '@/presentation/api/impl/ZodLocationPresenter.js'
 import { BuildingInsertion, RoomInsertion, RoomUpdate } from '@/presentation/api/schemas/LocationSchemas'
+import { Location } from '@/domain/core/Location'
 
 export const router: Router = express.Router()
 const locationPresenter: LocationPresenter = new ZodLocationPresenter()
@@ -12,7 +13,7 @@ router
   .route('/:id')
   .get(async (req: Request, res: Response) => {
     try {
-      const location = await locationsController.getLocationById(req.params.id)
+      const location: Location = await locationsController.getLocationById(req.params.id)
       locationPresenter.parse(location)
       res.status(HttpStatusCode.OK).json(location)
     } catch (e) {
@@ -32,7 +33,7 @@ router
   .route('/rooms')
   .get(async (req: Request, res: Response) => {
     try {
-      const rooms = await locationsController.getRooms()
+      const rooms: Location[] = await locationsController.getRooms()
       res.status(HttpStatusCode.OK).json(rooms)
     } catch (error) {
       res.status(HttpStatusCode.BAD_REQUEST).json({ message: 'error' })
@@ -62,7 +63,7 @@ router
   .route('/buildings')
   .get(async (req: Request, res: Response) => {
     try {
-      const buildings = await locationsController.getBuildings()
+      const buildings: Location[] = await locationsController.getBuildings()
       buildings.forEach(building => {
         locationPresenter.parse(building)
       })
@@ -93,7 +94,7 @@ router.route('/buildings/:id').put(async (req: Request, res: Response) => {
 
 router.route('/buildings/:buildingId/rooms').get(async (req: Request, res: Response) => {
   try {
-    const rooms = await locationsController.getBuildingRooms(req.params.buildingId)
+    const rooms: Location[] = await locationsController.getBuildingRooms(req.params.buildingId)
     rooms.forEach(room => {
       locationPresenter.parse(room)
     })
